requirements.txt
netmiko
ncclient
pyyaml
